Refactor: move IntoNodes into separate file
