+++ openraft/src/membership/mod.rs
@@ -1,5 +1,6 @@
 mod change_handler;
 mod effective_membership;
+mod into_nodes;
 #[allow(clippy::module_inception)] mod membership;
 mod membership_state;
 mod stored_membership;
@@ -14,7 +15,7 @@ mod bench;
 
 pub(crate) use change_handler::ChangeHandler;
 pub use effective_membership::EffectiveMembership;
-pub use membership::IntoNodes;
+pub use into_nodes::IntoNodes;
 pub use membership::Membership;
 pub use membership_state::MembershipState;
 pub use stored_membership::StoredMembership;
